+++ src/ska_mid_dish_manager/component_managers/dish_manager_cm.py
@@ -81,8 +81,6 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             desiredpointingaz=[0.0, 0.0],
             desiredpointingel=[0.0, 0.0],
             achievedpointing=[0.0, 0.0, 0.0],
-            achievedpointingaz=[0.0, 0.0, 0.0],
-            achievedpointingel=[0.0, 0.0, 0.0],
             configuredband=Band.NONE,
             attenuationpolh=0.0,
             attenuationpolv=0.0,
@@ -147,8 +145,6 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 desiredpointingaz=[0.0, 0.0],
                 desiredpointingel=[0.0, 0.0],
                 achievedpointing=[0.0, 0.0, 0.0],
-                achievedpointingaz=[0.0, 0.0, 0.0],
-                achievedpointingel=[0.0, 0.0, 0.0],
                 band1pointingmodelparams=[],
                 band2pointingmodelparams=[],
                 band3pointingmodelparams=[],
@@ -224,8 +220,6 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         self.direct_mapped_attrs = {
             "DS": [
                 "achievedPointing",
-                "achievedPointingAz",
-                "achievedPointingEl",
                 "desiredPointingAz",
                 "desiredPointingEl",
                 "trackInterpolationMode",